requirements.txt
Flask==2.3.3
requests==2.31.0
debugpy # Required for debugging.
gunicorn==21.2.0
pandas==2.2.2
openpyxl==3.1.5
google-auth==2.35.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography
